fix: update package name to com.swent.mapin
