+++ app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:816281112017:android:fb9cdf95cd29774171e4f0",
         "android_client_info": {
-          "package_name": "com.mapin"
+          "package_name": "com.swent.mapin"
         }
       },
       "oauth_client": [],
