+++ app/src/main/java/com/swent/mapin/resources/C.kt
@@ -1,4 +1,4 @@
-package com.android.sample.resources
+package com.swent.mapin.resources
 
 // Like R, but C
 object C {
