+++ app/src/test/java/com/swent/mapin/PointTest.kt
@@ -1,4 +1,4 @@
-package com.android.sample
+package com.swent.mapin
 
 import org.junit.Assert.assertEquals
 import org.junit.Test
